requirements.txt
streamlit
google-generativeai
requests
pydantic
python-dotenv

